🚚 Post 페이지 컴포넌트 파파일 구조 정리
