+++ src/pages/post/components/HeaderFeatureBox/HeaderFeatureBox.jsx
@@ -1,5 +1,5 @@
-import MessageRecipient from "./MessageRecipient";
-import HeaderServiceBox from "./HeaderServiceBox";
+import MessageRecipient from "../MessageRecipient/MessageRecipient";
+import HeaderServiceBox from "../HeaderServiceBox/HeaderServiceBox";
 import * as S from "./HeaderFeatureBox.style";
 const HeaderFeatureBox = () => {
   return (
