+++ src/pages/post/components/HeaderFeatureBox/HeaderFeatureBox.style.jsx
