+++ src/pages/post/components/HeaderLogoBox/HeaderLogoBox.style.jsx
