+++ src/pages/post/components/HeaderServiceBox/HeaderServiceBox.jsx
@@ -1,7 +1,7 @@
 import * as S from "./HeaderServiceBox.style";
-import MessageWriterBox from "./MessageWriterBox";
-import MostEmojiBox from "./MostEmojiBox";
-import HeaderButtonBox from "./HeaderButtonBox";
+import MessageWriterBox from "../MessageWriterBox/MessageWriterBox";
+import MostEmojiBox from "../MostEmojiBox/MostEmojiBox";
+import HeaderButtonBox from "../HeaderButtonBox/HeaderButtonBox";
 
 const HeaderServiceBox = () => {
   return (
