+++ src/pages/post/components/PostPageMain/PostPageMain.style.jsx
