+++ src/pages/post/components/WriterImageBox/WriterImageBox.style.jsx
