+++ src/pages/post/index.jsx
@@ -1,8 +1,8 @@
 import * as S from "./index.style";
 
-import PostPageMain from "./components/PostPageMain";
-import HeaderLogoBox from "./components/HeaderLogoBox";
-import HeaderFeatureBox from "./components/HeaderFeatureBox";
+import PostPageMain from "./components/PostPageMain/PostPageMain";
+import HeaderLogoBox from "./components/HeaderLogoBox/HeaderLogoBox";
+import HeaderFeatureBox from "./components/HeaderFeatureBox/HeaderFeatureBox";
 
 const PostPage = () => {
   return (
